Add Bulk Article API

+++ portality/api/v1/crud/articles.py
@@ -54,7 +54,7 @@ class ArticlesCrudApi(CrudApi):
         return cls._build_swag_response(template)
 
     @classmethod
-    def create(cls, data, account):
+    def create(cls, data, account, dry_run=False):
         # as long as authentication (in the layer above) has been successful, and the account exists, then
         # we are good to proceed
         if account is None:
@@ -73,17 +73,25 @@ class ArticlesCrudApi(CrudApi):
         if not XWalk.is_legitimate_owner(am, account.id):
             raise Api403Error()
 
-        # if the caller set the id, created_date, or last_updated, then we discard the data and apply our
-        # own values (note that last_updated will get overwritten anyway)
-        am.set_id()
-        am.set_created()
+        # before finalising, we need to determine whether this is a new article
+        # or an update
+        duplicate = XWalk.get_duplicate(am, account.id)
+        # print duplicate
+        if duplicate is not None:
+            am.merge(duplicate) # merge will take the old id, so this will overwrite
+        else:
+            # if the caller set the id, created_date, or last_updated, then we discard the data and apply our
+            # own values (note that last_updated will get overwritten anyway)
+            am.set_id()
+            am.set_created()
 
         # not allowed to set subjects
         am.bibjson().remove_subjects()
         am = cls.__handle_journal_info(am)
 
         # finally save the new article, and return to the caller
-        am.save()
+        if not dry_run:
+            am.save()
         return am
 
     @classmethod
@@ -181,7 +189,7 @@ class ArticlesCrudApi(CrudApi):
         return cls._build_swag_response(template)
 
     @classmethod
-    def delete(cls, id, account):
+    def delete(cls, id, account, dry_run=False):
         # as long as authentication (in the layer above) has been successful, and the account exists, then
         # we are good to proceed
         if account is None:
@@ -197,4 +205,5 @@ class ArticlesCrudApi(CrudApi):
             raise Api404Error()  # not found for this account
 
         # issue the delete (no record of the delete required)
-        ar.delete()
+        if not dry_run:
+            ar.delete()
